refactor(View): Impl Basic components View of UI
