+++ src/view/help.rs
@@ -4,6 +4,10 @@ use tui::{
     text::{Span, Spans},
 };
 
+
+// -----------------------------------------------------------------
+// OLD        ------------------------------------------------------
+// -----------------------------------------------------------------
 #[derive(Deserialize, Serialize, Clone)]
 enum StyleOptions {
     ColorCyan,
