+++ src/view/mod.rs
@@ -3,6 +3,10 @@ pub mod ui;
 pub mod renderer;
 pub mod components;
 
+
+// -----------------------------------------------------------------
+// OLD      --------------------------------------------------------
+// -----------------------------------------------------------------
 use crate::base::states::names::StatesNames;
 use crate::base::stores::MainStore;
 use tui::{
